+++ crates/aptos-metrics-core/src/avg_counter.rs
@@ -1,61 +1,15 @@
 // Copyright © Aptos Foundation
 
-use prometheus::{register_counter, register_int_counter, Counter, IntCounter};
+use prometheus::{register_histogram, Histogram};
 
-pub struct AverageCounter {
-    sum: Counter,
-    count: IntCounter,
-}
-
-impl AverageCounter {
-    pub fn register(name: &str, desc: &str) -> AverageCounter {
-        AverageCounter {
-            sum: register_counter!(
-                format!("{}_sum", name),
-                format!("{}. Sum part of the counter", desc),
-            )
-            .unwrap(),
-            count: register_int_counter!(
-                format!("{}_count", name),
-                format!("{}. Count part of the counter", desc),
-            )
-            .unwrap(),
-        }
-    }
-
-    pub fn observe(&self, value: f64) {
-        if value != 0.0 {
-            self.sum.inc_by(value);
-        }
-        self.count.inc();
-    }
-}
-
-pub struct AverageIntCounter {
-    sum: IntCounter,
-    count: IntCounter,
-}
-
-impl AverageIntCounter {
-    pub fn register(name: &str, desc: &str) -> AverageIntCounter {
-        AverageIntCounter {
-            sum: register_int_counter!(
-                format!("{}_sum", name),
-                format!("{}. Sum part of the counter", desc),
-            )
-            .unwrap(),
-            count: register_int_counter!(
-                format!("{}_count", name),
-                format!("{}. Count part of the counter", desc),
-            )
-            .unwrap(),
-        }
-    }
-
-    pub fn observe(&self, value: u64) {
-        if value != 0 {
-            self.sum.inc_by(value);
-        }
-        self.count.inc();
-    }
+// use histogram, instead of pair of sum/count counters, to guarantee
+// atomicity of observing and fetching (which Histogram handles correctly)
+pub fn register_avg_counter(name: &str, desc: &str) -> Histogram {
+    register_histogram!(
+        name,
+        desc,
+        // We need to have at least one bucket in histogram, otherwise default buckets are used
+        vec![0.5],
+    )
+    .unwrap()
 }