+++ docker/experimental/build-tools.sh
@@ -5,8 +5,9 @@ set -e
 
 PROFILE=${PROFILE:-release}
 
-echo "Building all rust-based docker images"
+echo "Building tools and services docker images"
 echo "PROFILE: $PROFILE"
+echo "CARGO_TARGET_DIR: $CARGO_TARGET_DIR"
 
 # Build all the rust binaries
 cargo build --locked --profile=$PROFILE \
@@ -19,7 +20,12 @@ cargo build --locked --profile=$PROFILE \
     -p aptos-openapi-spec-generator \
     -p aptos-telemetry-service \
     -p aptos-db-bootstrapper \
+    -p aptos-db-tool \
     -p aptos-transaction-emitter \
+    -p aptos-indexer-grpc-cache-worker \
+    -p aptos-indexer-grpc-file-store \
+    -p aptos-indexer-grpc-data-service \
+    -p aptos-indexer-grpc-parser \
     "$@"
 
 # After building, copy the binaries we need to `dist` since the `target` directory is used as docker cache mount and only available during the RUN step
@@ -30,19 +36,22 @@ BINS=(
     aptos-openapi-spec-generator
     aptos-telemetry-service
     aptos-fn-check-client
-    db-backup
-    db-backup-verify
+    aptos-db-tool
     aptos-db-bootstrapper
-    db-restore
     forge
     aptos-transaction-emitter
+    aptos-indexer-grpc-cache-worker
+    aptos-indexer-grpc-file-store
+    aptos-indexer-grpc-data-service
+    aptos-indexer-grpc-parser
 )
 
 mkdir dist
 
 for BIN in "${BINS[@]}"; do
-    cp target/$PROFILE/$BIN dist/$BIN
+    cp $CARGO_TARGET_DIR/$PROFILE/$BIN dist/$BIN
 done
 
 # Build the Aptos Move framework and place it in dist. It can be found afterwards in the current directory.
+echo "Building the Aptos Move framework..."
 (cd dist && cargo run --locked --profile=$PROFILE --package aptos-framework -- release)