+++ ecosystem/typescript/sdk/src/tests/e2e/provider.test.ts
@@ -5,13 +5,13 @@ import { Provider } from "../../providers/provider";
 import { FaucetClient } from "../../plugins/faucet_client";
 import { TokenClient } from "../../plugins/token_client";
 import { Network, NetworkToIndexerAPI, NetworkToNodeAPI, sleep } from "../../utils";
-import { API_TOKEN, longTestTimeout } from "../unit/test_helper.test";
+import { FAUCET_AUTH_TOKEN, longTestTimeout } from "../unit/test_helper.test";
 
 describe("Provider", () => {
   const faucetClient = new FaucetClient(
     "https://fullnode.testnet.aptoslabs.com",
     "https://faucet.testnet.aptoslabs.com",
-    { TOKEN: API_TOKEN },
+    { TOKEN: FAUCET_AUTH_TOKEN },
   );
   const alice = new AptosAccount();